pyproject.toml
[tool.poetry]
name = "ebi-ngff-challenge"
version = "0.1.0"
description = ""
authors = ["Matthew Hartley <matthewh@ebi.ac.uk>"]
readme = "README.md"
packages = [{include = "python_example"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
bia-integrator-api = { git = "https://github.com/BioImage-Archive/bia-integrator.git", subdirectory = "clients/python", branch = "main" }
rich = "^13.8.0"
ome2024-ngff-challenge = "^1.0.2"
click = "^8.1.7"
typer = "^0.12.5"
ruamel-yaml = "^0.18.6"
pydantic-settings = "^2.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
